requirements.txt
# XAI Lung Segmentation Analysis - Python Dependencies

# ============================================================================
# Core Deep Learning & ML
# ============================================================================
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# ============================================================================
# Data Science & Numerical Computing
# ============================================================================
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0
scikit-learn>=1.0.0
scikit-image>=0.19.0

# ============================================================================
# Visualization & Plotting
# ============================================================================
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0
opencv-python>=4.5.0

# ============================================================================
# Interactive Dashboard
# ============================================================================
streamlit>=1.25.0
streamlit-plotly-events>=0.0.6

# ============================================================================
# Image Processing
# ============================================================================
Pillow>=8.3.0
imageio>=2.9.0
albumentations>=1.0.0

# ============================================================================
# Configuration & Data Handling
# ============================================================================
PyYAML>=6.0
pydantic>=1.8.0
omegaconf>=2.1.0

# ============================================================================
# XAI & Interpretability
# ============================================================================
captum>=0.6.0
lime>=0.2.0
shap>=0.41.0

# ============================================================================
# Progress & Monitoring
# ============================================================================
tqdm>=4.62.0
tensorboard>=2.8.0
wandb>=0.12.0

# ============================================================================
# Development & Testing
# ============================================================================
pytest>=6.2.0
pytest-cov>=3.0.0
black>=22.0.0
flake8>=4.0.0
mypy>=0.910
pre-commit>=2.15.0

# ============================================================================
# Documentation
# ============================================================================
sphinx>=4.0.0
sphinx-rtd-theme>=1.0.0
myst-parser>=0.17.0

# ============================================================================
# Utilities
# ============================================================================
click>=8.0.0
requests>=2.25.0
python-dotenv>=0.19.0
rich>=12.0.0

# ============================================================================
# Medical Imaging (Optional)
# ============================================================================
nibabel>=3.2.0
pydicom>=2.2.0
SimpleITK>=2.1.0

# ============================================================================
# Performance & Optimization
# ============================================================================
numba>=0.56.0
joblib>=1.1.0